updated class card and popup for it to use new colors
