+++ src/pages/dashboard/InstructorDashboard.tsx
@@ -6,11 +6,28 @@ import LineChart from "../../components/LineChart";
 import { useInstructorClasses } from "../../hooks/useInstructorClasses";
 import StackedBarChart from "../../components/StackedBarChart";
 import CreateNewClassDialog from "../../components/CreateNewClassDialog";
-import { User, UserActivityLogItem } from "../../api/types/user";
+import {
+  ClassData,
+  StudentStatus,
+  User,
+  UserActivityLogItem,
+  UserRole,
+} from "../../api/types/user";
 import { ProgressData } from "../../utils/calculateProgress";
 import { Card } from "../../components/ui/card";
 import InfoTooltip from "../../components/InfoTooltip";
 import StudentDataTable from "../../components/StudentDataTable";
+import {
+  Carousel,
+  CarouselContent,
+  CarouselItem,
+  CarouselNext,
+  CarouselPrevious,
+} from "../../components/ui/carousel";
+import ClassInfoCard from "../../components/ClassInfoCard";
+import ClassDetails from "../../components/ClassDetails";
+import { useState } from "react";
+import { Loader2 } from "lucide-react";
 
 /**
  * InstructorDashboard component displays the instructor dashboard with class statistics and activity logs.
@@ -154,5 +171,96 @@ export const InstructorStudents = ({ userData }: { userData: User }) => {
         classFilter={selectedClassTitle === "all classes" ? "all" : "class"}
       />
     </Card>
+  );
+};
+
+export const InstructorClasses = ({ userData }: { userData: User }) => {
+  const { originalClasses, loading: userClassLoading } = useInstructorClasses();
+  const [selectedClass, setSelectedClass] = useState<{
+    userClass: ClassData;
+    studentStatus?: StudentStatus;
+    instructorData?: User;
+  } | null>(null);
+
+  const handleCloseDetails = () => {
+    setSelectedClass(null);
+  };
+  const handleClassSelect = (
+    userClass: ClassData,
+    studentStatus?: StudentStatus,
+    instructorData?: User
+  ) => {
+    setSelectedClass({
+      userClass,
+      studentStatus,
+      instructorData,
+    });
+  };
+
+  if (userClassLoading) {
+    return (
+      <div className="flex min-h-screen justify-center items-center">
+        <Loader2 className="size-12 animate-spin" />
+      </div>
+    );
+  }
+
+  return (
+    <div className="width-container grid grid-cols-1 md:grid-cols-5 lg:grid-cols-3 gap-6">
+      <div className="col-span-1 md:col-span-3 lg:col-span-2 space-y-4">
+        <div className="card flex w-full justify-center">
+          {originalClasses.length === 0 ? (
+            <div className="flex flex-col items-center justify-center gap-2 p-16 text-center">
+              <p className="text-lg font-medium">
+                You currently have no classes
+              </p>
+              <p className="text-muted-foreground mb-6">
+                Please create a new class
+              </p>
+              <CreateNewClassDialog />
+            </div>
+          ) : (
+            <Carousel
+              opts={{
+                align: "start",
+              }}
+              className="w-full"
+            >
+              <CarouselContent>
+                {originalClasses.map((userClass, index) => (
+                  <CarouselItem key={index} className="lg:basis-1/2">
+                    <div className="p-1">
+                      <ClassInfoCard
+                        classInfo={userClass}
+                        onSelect={handleClassSelect}
+                        isInstructor={
+                          userData.role === UserRole.INSTRUCTOR ||
+                          userData.role === UserRole.ADMIN
+                        }
+                      />
+                    </div>
+                  </CarouselItem>
+                ))}
+              </CarouselContent>
+              {originalClasses.length > 1 && (
+                <div
+                  className={`flex justify-between w-full ${originalClasses.length <= 2 && "lg:hidden"}`}
+                >
+                  <CarouselPrevious className="ml-4" />
+                  <CarouselNext className="mr-4" />
+                </div>
+              )}
+            </Carousel>
+          )}
+        </div>
+      </div>
+      {selectedClass && (
+        <ClassDetails
+          userClass={selectedClass.userClass}
+          instructorData={selectedClass.instructorData as User}
+          onClose={handleCloseDetails}
+        />
+      )}
+    </div>
   );
 };
